CONF: make conftest.py in common directory
